latest push on IndexMonitor work. Pushing to get feedback from @sbcd90.
