+++ src/main/java/org/opensearch/securityanalytics/alerting/resthandlers/RestIndexMonitorAction.java
@@ -7,19 +7,24 @@ package org.opensearch.securityanalytics.alerting.resthandlers;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.json.JSONObject;
 import org.opensearch.action.support.WriteRequest;
 import org.opensearch.client.node.NodeClient;
+import org.opensearch.commons.model2.ModelSerializer;
 import org.opensearch.commons.model2.action.IndexMonitorAction;
 import org.opensearch.commons.model2.action.IndexMonitorRequest;
 import org.opensearch.commons.model2.model.Monitor;
+import org.opensearch.index.seqno.SequenceNumbers;
 import org.opensearch.rest.BaseRestHandler;
 import org.opensearch.rest.RestRequest;
 import org.opensearch.rest.action.RestStatusToXContentListener;
 
 import java.io.IOException;
+import java.nio.charset.Charset;
 import java.util.List;
 import java.util.Set;
 
+import static org.opensearch.commons.model2.action.IndexMonitorResponse.*;
 import static org.opensearch.rest.RestRequest.Method.POST;
 import static org.opensearch.securityanalytics.Tokens.*;
 
@@ -47,15 +52,15 @@ public class RestIndexMonitorAction extends BaseRestHandler {
             LOG.info("{} {}/{}", request.method(), SAP_MONITORS_BASE_URI, getName());
             final String monitorId = request.param(MONITOR_ID);
             if (null == monitorId || monitorId.isEmpty()) throw new IllegalArgumentException("missing monitorID");
-            /*long seqNo = request.paramAsLong(IF_SEQ_NO, SequenceNumbers.UNASSIGNED_SEQ_NO);
+            long seqNo = request.paramAsLong(IF_SEQ_NO, SequenceNumbers.UNASSIGNED_SEQ_NO);
             long primaryTerm = request.paramAsLong(IF_PRIMARY_TERM, SequenceNumbers.UNASSIGNED_PRIMARY_TERM);
             WriteRequest.RefreshPolicy refreshPolicy = (request.hasParam(REFRESH)) ? WriteRequest.RefreshPolicy.parse(request.param(REFRESH)) : WriteRequest.RefreshPolicy.IMMEDIATE;
-            XContentParser parser = request.contentParser();
-            final Monitor monitor = ModelSerializer.read(parser, Monitor.class);*/
-
+            // TODO: I simply turned the POST body into a JSONObject (nested map) and used ModelSerializer to generate the monitor
+            final JSONObject json = new JSONObject(new String(request.content().toBytesRef().bytes, Charset.defaultCharset()));
+            final Monitor monitor = ModelSerializer.read(json, Monitor.class);
             client.execute(
-                    IndexMonitorAction.SAP_ALERTING_BRIDGE_INSTANCE,
-                    new IndexMonitorRequest(monitorId, 3L, 4L, WriteRequest.RefreshPolicy.IMMEDIATE, request.method(), new Monitor("anid", "atype", 12L, null, "hello", 235L, "seconds", List.of())),
+                    IndexMonitorAction.SAP_INSTANCE,
+                    new IndexMonitorRequest(monitorId, seqNo, primaryTerm, refreshPolicy, request.method(), monitor),
                     new RestStatusToXContentListener(channel));
         };
     }
